+++ packages/core/src/data/strategies/SearchNativeFetchStrategy.ts
@@ -141,7 +141,7 @@ export class SearchNativeFetchStrategy<
 					addQueryArgs(`${wpUrl}${endpoints.yoast}`, {
 						url: `${wpUrl}${pageParam}/?s=${params.search ?? ''}${localeParam}`,
 					}),
-					{ headers: options.headers ?? {} },
+					{ headers: options.headers ?? {}, cache: options?.cache },
 					burstCache,
 				);