+++ packages/next/src/rsc/data/queries/queryAuthorArchive.ts
@@ -1,4 +1,5 @@
 import {
+	AuthorArchiveFetchStrategy,
 	HeadlessConfig,
 	PostEntity,
 	PostsArchiveParams,
@@ -9,17 +10,26 @@ import { NextQueryProps } from './types';
 import { prepareQuery } from './prepareQuery';
 import { prepareSEOMetadata } from '../seo';
 
+export type AuthorArchiveQueryProps<
+	T extends PostEntity = PostEntity,
+	P extends PostsArchiveParams = PostsArchiveParams,
+> = NextQueryProps<P> & {
+	fetchStrategy?: AuthorArchiveFetchStrategy<T, P>;
+};
+
 export async function queryAuthorArchive<
 	T extends PostEntity = PostEntity,
 	P extends PostsArchiveParams = PostsArchiveParams,
->(q: NextQueryProps<P> = {}, _config: HeadlessConfig | undefined = undefined) {
-	const { config, handleError, ...query } = prepareQuery<P>(q, _config);
+>(q: AuthorArchiveQueryProps<T, P> = {}, _config: HeadlessConfig | undefined = undefined) {
+	const { fetchStrategy, ...nextQuery } = q;
+	const { config, handleError, ...query } = prepareQuery<P>(nextQuery, _config);
 
 	try {
-		const result = await fetchAuthorArchive<T, P>(query, config);
+		const result = await fetchAuthorArchive<T, P>(query, config, fetchStrategy);
 
 		return {
 			...result,
+			config,
 			seo: prepareSEOMetadata(result.data.queriedObject, config),
 		};
 	} catch (error) {