+++ packages/next/src/rsc/index.ts
@@ -9,3 +9,10 @@ export * from './handlers/revalidateRouterHandler';
 // components
 export * from './components/PreviewIndicator';
 export * from './components/JSONLD';
+export * from './components/Link';
+export * from './components/HeadstartWPApp';
+
+// blocks
+export * from './blocks/LinkBlock';
+export * from '../blocks/TwitterBlock';
+export * from './blocks/ImageBlock';