+++ projects/wp-nextjs-app/src/app/(single)/[...path]/page.tsx
@@ -1,7 +1,41 @@
 import { HtmlDecoder } from '@headstartwp/core/react';
-import { HeadstartWPRoute, JSONLD, queryPost } from '@headstartwp/next/app';
+import {
+	HeadstartWPRoute,
+	JSONLD,
+	loadHeadstartWPConfig,
+	queryPost,
+	queryPosts,
+} from '@headstartwp/next/app';
 import { Metadata } from 'next';
+import { removeSourceUrl } from '@headstartwp/core';
+import dynamic from 'next/dynamic';
+import { Suspense } from 'react';
 import Blocks from '../../../components/Blocks';
+import { ServerRelatedPosts } from '../../../components/ServerRelatedPosts';
+
+const ClientRelatedPosts = dynamic(() =>
+	import('../../../components/RelatedPosts').then((mod) => mod.RelatedPosts),
+);
+
+export async function generateStaticParams({ params }: HeadstartWPRoute) {
+	// loads the right config based on route params (this is needed over getHeadstartWP for sites using polylang integration or multisite)
+	const { sourceUrl = '', hostUrl = '/' } = loadHeadstartWPConfig(params);
+
+	const { data } = await queryPosts({
+		routeParams: params,
+		params: { postType: 'post' },
+	});
+
+	return data.posts.map((post) => ({
+		path: removeSourceUrl({
+			backendUrl: sourceUrl,
+			link: post.link,
+			publicUrl: hostUrl,
+		})
+			.substring(1)
+			.split('/'),
+	}));
+}
 
 async function query({ params }: HeadstartWPRoute) {
 	return queryPost({
@@ -21,7 +55,7 @@ export async function generateMetadata({ params }: HeadstartWPRoute): Promise<Me
 }
 
 const Single = async ({ params }: HeadstartWPRoute) => {
-	const { data, seo } = await query({ params });
+	const { data, seo, config } = await query({ params });
 
 	return (
 		<article>
@@ -29,9 +63,24 @@ const Single = async ({ params }: HeadstartWPRoute) => {
 				<HtmlDecoder html={data.post.title.rendered ?? ''} />
 			</h1>
 
-			<Blocks html={data.post.content.rendered ?? ''} />
+			<Blocks html={data.post.content.rendered ?? ''} settings={config} />
 
 			{seo.schema && <JSONLD schema={seo.schema} />}
+
+			{data.post.terms.category && (
+				<>
+					<ClientRelatedPosts
+						post_id={data.post.id}
+						category={data.post.terms.category[0].slug}
+					/>
+					<Suspense fallback="Loading (streaming)">
+						<ServerRelatedPosts
+							post_id={data.post.id}
+							category={data.post.terms.category[0].slug}
+						/>
+					</Suspense>
+				</>
+			)}
 		</article>
 	);
 };