+++ projects/wp-nextjs-app/src/app/api/revalidate/route.ts
@@ -1,7 +1,8 @@
 import { revalidateRouteHandler } from '@headstartwp/next/app';
 import type { NextRequest } from 'next/server';
 
+export const dynamic = 'force-dynamic';
+
 export async function GET(request: NextRequest) {
-	// @ts-expect-error
 	return revalidateRouteHandler(request);
 }