+++ projects/wp-nextjs/src/pages/_app.tsx
@@ -41,7 +41,6 @@ const MyApp = ({ Component, pageProps }: AppProps<MyAppProps>) => {
 				revalidateOnMount: false,
 			}}
 			settings={{
-				// @ts-expect-error
 				linkComponent: Link,
 			}}
 			useYoastHtml