+++ packages/next/CHANGELOG.md
@@ -1,5 +1,77 @@
 # @headstartwp/next
 
+## 1.5.0-next.4
+
+### Patch Changes
+
+- 9ea83c3: Fix: return config from queryPostOrPosts
+- c1d70ec: Opt into image.remotePatters for Next.js versions that supports it
+- Updated dependencies [9ea83c3]
+  - @headstartwp/core@1.5.0-next.4
+
+## 1.5.0-next.3
+
+### Patch Changes
+
+- 0c735b7: Fixed - Fixed isValidLocale function to validate against the configuration
+  Added - New test condition for technically valid, but unsupported locale
+- Updated dependencies [fe250b7]
+  - @headstartwp/core@1.5.0-next.3
+
+## 1.5.0-next.2
+
+### Patch Changes
+
+- d49ac5c: Allow overriding fetch strategy
+- Updated dependencies [d49ac5c]
+  - @headstartwp/core@1.5.0-next.2
+
+## 1.5.0-next.1
+
+### Minor Changes
+
+- 915c6b6: RSC compatible image block for next.js
+
+### Patch Changes
+
+- Updated dependencies [915c6b6]
+  - @headstartwp/core@1.5.0-next.1
+
+## 1.5.0-next.0
+
+### Minor Changes
+
+- 6fa9ec2: Introducing previewRouteHandler to handle preview in Next.js App Router
+- 96ea386: Initial App Router Support
+- e969f38: Introducing SEO handling for App Router
+- 2c51bfb: Link handling
+- 79ae20e: Adding support for `i18n` routing in app router
+- 6fa9ec2: Introducing `revalidateRouteHandler` for handling revalidate requests in Route Handlers (App Router)
+- 4f7a665: Add `forwardBlockProps` to BlocksRenderer which automatically forwards block props to children components
+- ca94535: Introduce `blockContext` prop to BlocksRenderer and `handleError` in query functions
+- d696703: Multisite support in App Router
+
+### Patch Changes
+
+- 14c398d: remove usage of defaultProps and intropduce the ability to attach a default test function directly to the component
+- 4044248: Introducing `loadHeadstartWpConfig()`
+- 9f05f3f: Fix fetch cache option
+- c73d000: Fix: SEO with yoast disabled
+- Updated dependencies [6fa9ec2]
+- Updated dependencies [96ea386]
+- Updated dependencies [e969f38]
+- Updated dependencies [2c51bfb]
+- Updated dependencies [79ae20e]
+- Updated dependencies [14c398d]
+- Updated dependencies [6fa9ec2]
+- Updated dependencies [4f7a665]
+- Updated dependencies [4044248]
+- Updated dependencies [9f05f3f]
+- Updated dependencies [ca94535]
+- Updated dependencies [d696703]
+- Updated dependencies [c73d000]
+  - @headstartwp/core@1.5.0-next.0
+
 ## 1.4.3
 
 ### Patch Changes